Add mypy and fix mypy inspections

+++ malcolm/modules/adUtil/parts/reframepluginpart.py
@@ -1,4 +1,6 @@
-from annotypes import Anno, Any, add_call_types
+from typing import Any
+
+from annotypes import Anno, add_call_types
 
 from malcolm.core import Context, PartRegistrar
 from malcolm.modules import ADCore, builtin, scanning
@@ -39,7 +41,7 @@ class ReframePluginPart(ADCore.parts.DetectorDriverPart):
         context: Context,
         completed_steps: scanning.hooks.ACompletedSteps,
         steps_to_do: scanning.hooks.AStepsToDo,
-        duration: int,
+        duration: float,
         part_info: scanning.hooks.APartInfo,
         **kwargs: Any,
     ) -> None:
